bothack/bot/requirements.txt
python-telegram-bot
SQLAlchemy
psycopg2-binary
aiohttp